+++ plone/app/event/base.py
@@ -32,6 +32,8 @@ class CalendarLinkbase(object):
     """Default adapter to retrieve a base url for a calendar view. The methods
     in this default implementation return the @@search view as calendar view.
 
+    For method documentation, see interfaces.py.
+
     """
     adapts(Interface)
     implements(ICalendarLinkbase)
@@ -39,104 +41,29 @@ class CalendarLinkbase(object):
     def __init__(self, context):
         self.context = context
         portal = getSite()
-        self.navroot = getNavigationRootObject(context, portal)
-        self.navroot_url = self.navroot.absolute_url()
+        self.navroot = getNavigationRootObject(context, portal).absolute_url()
 
     def date_events_url(self, date, path=None):
-        """Get a URL to retrieve all events on a given day.
-
-        :param date: The date to search events for in isoformat (ISO 8601,
-                    YYYY-MM-DD).
-        :type date: string
-
-        :param path: Events context path (optional).
-        :type path: string
-
-        :returns: URL linking to a page with events on the given date.
-        :rtype: string
-
-        """
-        url = '%s/@@search?advanced_search=True&'\
-              'start.query:record:list:date=%s+23:59:59&'\
-              'start.range:record=max&'\
-              'end.query:record:list:date=%s+00:00:00&'\
-              'end.range:record=min&'\
-              'object_provides=plone.event.interfaces.IEvent'\
-              % (self.navroot_url, date, date)
+        url = '%s/@@event_listing?mode=day&date=%s' % (self.navroot, date)
         return url
 
     def past_events_url(self):
         """Get a URL to retrieve past events.
-
-        :returns: URL linking to a page with past events.
-        :rtype: string
-
         """
-        # take care dont use self.portal here since support
-        # of INavigationRoot features likely will breake #9246 #9668
-        url = None
-        navigation_root_url = self.navroot_url
-        events_folder = self._events_folder()
-        if (events_folder and
-            'aggregator' in events_folder.objectIds() and
-            'previous' in events_folder['aggregator'].objectIds()):
-            url = '%s/events/aggregator/previous' % navigation_root_url
-        elif (events_folder and 'previous' in events_folder.objectIds()):
-            url = '%s/events/previous' % navigation_root_url
-        else:
-            # show all past events
-            now = datetime.utcnow().strftime('%Y-%m-%d+%H:%M')
-            url = '%s/@@search?advanced_search=True'\
-                  '&end.query:record:list:date=%s'\
-                  '&end.range:record=max'\
-                  '&object_provides=plone.event.interfaces.IEvent'\
-                   % (navigation_root_url, now)
+        url = '%s/@@event_listing?mode=past' % (self.navroot)
         return url
 
     def next_events_url(self):
         """Get a URL to retrieve upcoming events.
-
-        :returns: URL linking to a page with upcoming events.
-        :rtype: string
-
         """
-        navigation_root_url = self.navroot_url
-        url = None
-        if self._events_folder():
-            url = '%s/events' % navigation_root_url
-        else:
-            # search all events which are in the future or ongoing
-            now = datetime.utcnow().strftime('%Y-%m-%d+%H:%M')
-            url = '%s/@@search?advanced_search=True'\
-                  '&start.query:record:list:date=%s'\
-                  '&start.range:record=min'\
-                  '&end.query:record:list:date=%s'\
-                  '&end.range:record=min'\
-                  '&object_provides=plone.event.interfaces.IEvent'\
-                   % (navigation_root_url, now, now)
+        url = '%s/@@event_listing?mode=future' % (self.navroot)
         return url
 
     def all_events_url(self):
         """Get a URL to retrieve all events.
-
-        :returns: URL linking to a page with events on the given date.
-        :rtype: string
-
         """
-        navigation_root_url = self.navroot_url
-        url = None
-        if self._events_folder():
-            url = '%s/events' % navigation_root_url
-        else:
-            # search all events which are in the future or ongoing
-            url = '%s/@@search?advanced_search=True'\
-                  '&object_provides=plone.event.interfaces.IEvent'\
-                   % navigation_root_url
+        url = '%s/@@event_listing?mode=all' % (self.navroot)
         return url
-
-    def _events_folder(self):
-        navroot = self.navroot
-        return 'events' in navroot and navroot['events'] or None
 
 
 def dt_start_of_day(dt):
@@ -164,27 +91,33 @@ def dt_end_of_day(dt):
     return dt.replace(hour=23, minute=59, second=59, microsecond=0)
 
 
-def start_end_from_mode(mode, context=None):
+def start_end_from_mode(mode, dt=None, context=None):
     """Return a start and end date from a given mode string, like
     "today", "past" or "future". This can be used in event retrieval
     functions.
 
-    :param mode: Optional. One of the following modes:
+    :param mode: One of the following modes:
                     'all' Show all events.
                     'past': Show only past events with descending sorting.
                     'future': Show only future events (default).
                     'today': Show todays events.
                     'now': Show todays upcoming events.
                     '7days': Show events until 7 days in future.
+                    'day': Return all events on the given day (dt parameter
+                           required)
                  These settings override the start and end parameters.
                     Not implemented yet:
                     'week': Show this weeks events.
                     'month': Show this month's events.
     :type mode: string
 
+    :param dt: Optional datetime for day mode.
+    :type dt: Python datetime
+
     """
     if not context: context = getSite()
     now = localized_now(context)
+    start = end = None
     if mode == 'all':
         start = None
         end = None
@@ -203,6 +136,9 @@ def start_end_from_mode(mode, context=None):
     elif mode == '7days':
         start = now
         end = dt_end_of_day(now+timedelta(days=7))
+    elif mode == 'day' and isinstance(dt, datetime):
+        start = dt_start_of_day(dt)
+        end = dt_end_of_day(dt)
 
     return start, end